Feat: Corrigido bugs e organização do código melhorada

+++ Back_End/Core/controllers/project_controller.py
@@ -9,38 +9,53 @@ router_project = APIRouter(
     tags=["Project"],
     responses={404: {"description": "Not found"}})
 
+
 @router_project.post("/create")
 async def create_project(createProject: Project_Create):
     project = await Project_Manager.create(
-        name = createProject.name,
-        description = createProject.description,
+        name = createProject.name, 
+        description = createProject.description, 
         status = createProject.status
         )
     return project
 
+
 @router_project.get("/getall")
 async def get_projects():
-    projetos = await Project_Manager.all()
-    return projetos
+    projects = await Project_Manager.all()
+    
+    return projects
 
 
 @router_project.post("/getone")
 async def get_project(Id: Get_Id):
-    projeto = await Project_Manager.get(id=Id.id)
-    return projeto
+    project = await Project_Manager.get_or_none(id=Id.id)
+    
+    if not project:
+        return {"Fail": True, "Result" : "Projeto não encontrado"}
+    
+    return {"Fail": False, "Result": project}
 
 
 @router_project.put("/update")
 async def update_project(Id: Get_Id, update_project: Project_Update):
-    project = await Project_Manager.get(id= Id.id)
+    project = await Project_Manager.get_or_none(id= Id.id)
+    
+    if not project:
+        return {"Fail": True, "Result": "Projeto não encontrado"}
+    
     project_update = update_project.model_dump(exclude_unset=True, exclude_defaults=True, exclude_none=True)
     project.update_from_dict(project_update)
     await project.save()
-    return project_update
+    return {"Fail": False, "Result" : project_update}
 
 
 @router_project.delete("/delete")
 async def delete_project(Id: Get_Id):
-    projeto = await Project_Manager.get(id=Id.id)
+    project = await Project_Manager.get_or_none(id=Id.id)
+    
+    if not project:
+        return {"Fail": True, "Result": "Projeto não encontrado"}
+    
     await Project_Manager.filter(id=Id.id).delete()
-    return projeto
+    return {"Fail": False, "Result": project}
